Deprecate Kiwi ConstraintValidator, make QualitativeConstraintValidator the default
Extract shared types (ConstraintError, PositionalConstraintError, etc.) and
the new IConstraintValidator interface into constraint-types.ts so they are
decoupled from the deprecated Kiwi-based validator.

Both validators now implement IConstraintValidator. The default strategy is
flipped from KIWI to QUALITATIVE, and the KIWI enum member is marked
@deprecated.

Also fixes a correctness bug in the qualitative validator where
findSatisfyingAlternative and isAlternativeFeasible did not check alignment
constraint alternatives against existing ordering relationships, causing
false-satisfiable results when a positive ordering constraint conflicted
with a negated orientation's alignment alternative.
