+++ src/layout/layoutinstance.ts
@@ -1,6 +1,6 @@
 import { Graph, Edge } from 'graphlib';
 import { IAtom, IDataInstance } from '../data-instance/interfaces';
-import { PositionalConstraintError, GroupOverlapError, isPositionalConstraintError, isGroupOverlapError, isHiddenNodeConflictError, HiddenNodeConflictError } from './constraint-validator';
+import { type PositionalConstraintError, type GroupOverlapError, type HiddenNodeConflictError, type IConstraintValidator, isPositionalConstraintError, isGroupOverlapError, isHiddenNodeConflictError } from './constraint-types';
 import { EdgeStyle, normalizeEdgeStyle } from './edge-style';
 import { resolveIconPath } from './icon-registry';
 import type { SelectorErrorDetail } from '../components/ErrorMessageModal/ErrorStateManager';
@@ -84,18 +84,18 @@ export enum AlignmentEdgeStrategy {
 /**
  * Strategy for which constraint validator to use.
  *
- * - `kiwi` (default): The original Kiwi-based backtracking validator.
- *   Mature and battle-tested.
+ * - `qualitative` (default): CDCL-based validator that reasons over qualitative
+ *   partial-order constraints before handing off to Kiwi for coordinate
+ *   assignment. Faster on large/group-heavy instances with geometry-aware
+ *   pruning and clause learning.
  *
- * - `qualitative` (beta): CDCL-based validator that reasons purely over
- *   qualitative partial-order constraints before handing off to Kiwi for
- *   coordinate assignment. Faster on large/group-heavy instances but still
- *   under active development.
+ * - `kiwi` (deprecated): The original Kiwi-based backtracking validator.
+ *   Retained for backward compatibility and equivalence testing.
  */
 export enum ConstraintValidatorStrategy {
-    /** Original Kiwi-based backtracking validator (default, stable) */
+    /** @deprecated Use QUALITATIVE instead. Will be removed in a future release. */
     KIWI = 'kiwi',
-    /** CDCL-based qualitative validator (beta) */
+    /** CDCL-based qualitative validator (default) */
     QUALITATIVE = 'qualitative'
 }
 
@@ -349,7 +349,7 @@ export class LayoutInstance {
      * @param instNum - The instance number (default is 0), used to differentiate between multiple instances of the same layout.
      * @param addAlignmentEdges - Deprecated. Use alignmentEdgeStrategy instead. A boolean flag indicating whether alignment edges should be added (default is `true`, equivalent to 'connected' strategy).
      * @param alignmentEdgeStrategy - Strategy for adding alignment edges (default is `AlignmentEdgeStrategy.CONNECTED`). Takes precedence over addAlignmentEdges if provided.
-     * @param validatorStrategy - Which constraint validator to use (default is `ConstraintValidatorStrategy.KIWI`). Set to `QUALITATIVE` to use the beta CDCL-based validator.
+     * @param validatorStrategy - Which constraint validator to use (default is `ConstraintValidatorStrategy.QUALITATIVE`). Set to `KIWI` to use the deprecated Cassowary-based validator.
      *
      * The `LayoutInstance` class is responsible for generating a layout for a given data instance based on the provided layout specification.
      * It applies constraints, directives, and projections to produce a structured layout that can be rendered using a graph visualization library.
@@ -376,7 +376,7 @@ export class LayoutInstance {
                 : AlignmentEdgeStrategy.NEVER;
         }
 
-        this.validatorStrategy = validatorStrategy ?? ConstraintValidatorStrategy.KIWI;
+        this.validatorStrategy = validatorStrategy ?? ConstraintValidatorStrategy.QUALITATIVE;
     }
 
     get hideDisconnected(): boolean {
@@ -1237,7 +1237,7 @@ export class LayoutInstance {
         }
 
         // Validate all constraints (conjunctive + disjunctive) in one pass
-        const validator = this.validatorStrategy === ConstraintValidatorStrategy.QUALITATIVE
+        const validator: IConstraintValidator = this.validatorStrategy === ConstraintValidatorStrategy.QUALITATIVE
             ? new QualitativeConstraintValidator(layout)
             : new ConstraintValidator(layout);
         const constraintError = validator.validateConstraints();
